Refactor objectTypes store
