+++ src/client/__tests__/reducers/fromObjectTypes.js
@@ -1,4 +1,7 @@
-import { getObjectTypesList, getObjectTypesLoading } from '../../store/reducers';
+import {
+  getObjectTypesList,
+  getObjectTypesLoading,
+} from '../../store/objectTypesStore/objectTypesSelectors';
 
 jest.mock('../../vendor/steemitHelpers.js', () => {});
