+++ src/client/discoverObjects/SidenavDiscoverObjects.js
@@ -4,9 +4,13 @@ import { useSelector, shallowEqual } from 'react-redux';
 import { NavLink } from 'react-router-dom';
 import { FormattedMessage, injectIntl } from 'react-intl';
 import { size, ceil } from 'lodash';
-import { getCurrentLocation, getObjectTypesList, getObjectTypesLoading } from '../store/reducers';
+import { getCurrentLocation } from '../store/reducers';
 import SkeletonCustom from '../components/Skeleton/SkeletonCustom';
 import { PATH_NAME_DISCOVER } from '../../common/constants/rewards';
+import {
+  getObjectTypesList,
+  getObjectTypesLoading,
+} from '../store/objectTypesStore/objectTypesSelectors';
 
 const typesLimit = 5;
 const SidenavDiscoverObjects = ({ withTitle, intl }) => {
